refactor: combine all enum code into a single file
